Ignore server-supplied "fs_cdn" value.
Kinda confusing, but here's what happens:

- On startup, the ioq3 client (web browser) is given a fs_cdn setting: an address to download assets from.
- When the client connects to the server, the server provides its own preferred cdn value as a part of the client connection string, I think?

Our server runs in Docker Compose and is told that its asset server is 'assets:9000', and this hostname doesn't resolve outside of docker.

In practice, 'assets:9000' inside docker compose is the same actual host as the web client is given, so it's safe to ignore.

If we do not ignore it, then the client attempts to download new pk3 files from "assets:9000" which fails either due to a timeout, cors failure, or other problem.